Rename the package `util` to `kotres`
For the consistency among Qoncept's Kotlin libraries.
