+++ test/jp/co/qoncept/kotres/Exceptions.kt
@@ -1,4 +1,4 @@
-package jp.co.qoncept.util
+package jp.co.qoncept.kotres
 
 open class AnimalException : Exception() {}
 class DogException : AnimalException() {}
